test: rename `common` module to `fixtures`
