+++ sec/src/lib/tests/fixtures/sample_http_client/sample_inner_client/always_failing.rs
